+++ src/tools/server/sdk/adGuard/adGuard.ts
@@ -4,7 +4,6 @@ import {
   adGuardApiStatsResponseSchema,
   adGuardApiStatusResponseSchema,
 } from './adGuard.schema';
-import { AdGuardStatsType } from './adGuard.type';
 
 export class AdGuard {
   private readonly baseHostName: string;
@@ -77,3 +76,20 @@ export class AdGuard {
     return Buffer.from(`${this.username}:${this.password}`).toString('base64');
   }
 }
+
+export type AdGuardStatsType = {
+  time_units: string;
+  top_queried_domains: { [key: string]: number }[];
+  top_clients: { [key: string]: number }[];
+  top_blocked_domains: { [key: string]: number }[];
+  dns_queries: number[];
+  blocked_filtering: number[];
+  replaced_safebrowsing: number[];
+  replaced_parental: number[];
+  num_dns_queries: number;
+  num_blocked_filtering: number;
+  num_replaced_safebrowsing: number;
+  num_replaced_safesearch: number;
+  num_replaced_parental: number;
+  avg_processing_time: number;
+};
